advanced Makefile, remove python source for notebooks, update deps
The new Makefile takes care of updating requirements.txt when installing a new
package specified in setup.cfg. It also uses the requirements.txt to create
populate the virtual environment if it is newer than the setup.cfg.

I have also removed the Python version of the notebooks for now, to simplify
the workflow (at the expense of less clean git diffs)